+++ example/lib/views/screens/trendings/content_page.dart
@@ -41,7 +41,7 @@ class ContentTrendings extends StatelessWidget {
                     child: Container(
                       padding: const EdgeInsets.all(4),
                       decoration: BoxDecoration(
-                          color: Theme.of(context).colorScheme.background,
+                          color: Theme.of(context).colorScheme.surface,
                           borderRadius: const BorderRadius.only(
                               topLeft: Radius.circular(10),
                               bottomRight: Radius.circular(10))),