+++ example/lib/widget/setting_group_card.dart
@@ -17,7 +17,7 @@ class SettingsGroup extends StatelessWidget {
     return Card(
       clipBehavior: Clip.antiAlias,
       elevation: 0,
-      color: Theme.of(context).colorScheme.background,
+      color: Theme.of(context).colorScheme.surface,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(12), // if you need this
         side: BorderSide(