+++ lib/models/channel_page.dart
@@ -1,8 +1,9 @@
 class ChannelPage {
   String? channelName;
   String? subscribers;
+  String? videoCounts;
   String? avatar;
   String? banner;
 
-  ChannelPage({this.channelName, this.subscribers, this.avatar, this.banner});
+  ChannelPage({this.channelName, this.subscribers, this.videoCounts, this.avatar, this.banner});
 }